refactor: 70.J.1.b — extract Buffer tab-stop methods to buffer/tabs.rs
Extract 6 tab-stop manipulation methods from buffer/mod.rs into a new
buffer/tabs.rs submodule (HT, HTS, TBC, CBT, CHT):

  - advance_to_next_tab_stop
  - set_tab_stop
  - clear_tab_stop_at_cursor
  - clear_all_tab_stops
  - tab_backward
  - tab_forward

Buffer struct fields are changed from private to pub(in crate::buffer)
so sibling modules inside buffer/ can access them. This is the minimum
privilege escalation: fields remain invisible outside the buffer/
module tree.

This is the first content-moving step of the 70.J.1 god-file split.
Subsequent commits will extract additional method groups (cursor,
scroll, erase, flatten, resize, images, alt-screen, text insert)
following the same pattern.

Verification: cargo test -p freminal-buffer green (all 500+ tests),
cargo clippy clean.
